pyproject.toml
[project]
name = "msg2po"
version = "1.4.3"
description = "A set of helper tools to convert Fallout 1/2 MSG and WeiDU TRA into GNU gettext PO and back."
authors = [{ name = "BGforge", email = "dev@bgforge.net" }]
keywords = ["Fallout", "Fallout 2", "Baldur's Gate", "Infinity Engine", "WeiDU"]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "natsort>=6.2.1",
    "polib>=1.1.1",
    "loguru>=0.7.0",
    "ruamel.yaml>=0.17.21",
]

[project.urls]
Homepage = "https://github.com/BGforgeNet/msg2po"
Repository = "https://github.com/BGforgeNet/msg2po"

[project.scripts]
"bgforge-config" = "msg2po.bgforge_config:main"
"dir2msgstr" = "msg2po.dir2msgstr:main"
"file2msgstr" = "msg2po.file2msgstr:main"
"file2po" = "msg2po.file2po:main"
"lowercase" = "msg2po.lowercase:main"
"msgmerge-female" = "msg2po.msgmerge:main"
"po2file" = "msg2po.po2file:main"
"poify" = "msg2po.poify:main"
"resave-po" = "msg2po.resave_po:main"
"unpoify" = "msg2po.unpoify:main"
"unfuzzy" = "msg2po.unfuzzy:main"

[dependency-groups]
dev = [
    "pre-commit>=3.5",
    "pytest>=7.4,<9",
    "ruff>=0.15",
    "ty>=0.0.1a8",
    "vulture>=2.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "SIM108", # ternary operator — often less readable in this codebase
]
